ml_app/.ipynb_checkpoints/requirements-checkpoint.txt
flask
numpy
matplotlib
tensorflow-cpu